requirements.txt
python-twitter-v2
kafka-python
python-dotenv
continuous-threading